MIPS: Don't re-sync r4k for hotplug
R4K is synced in SMP initialization, hotplug doesn't need to re-sync it.

And hotplug can not re-sync it for it may get a dead lock and make the
system hang.

	synchronise_count_master():

		[...]
		atomic_set(&count_count_start, 0);
		[...]

count_count_start is 0 after all cpus are up:

	synchronise_count_slave():

		[...]
	ncpus = num_online_cpus();
	for (i = 0; i < NR_LOOPS; i++) {
		atomic_inc(&count_count_start);
		while (atomic_read(&count_count_start) != ncpus)
			mb();
		[...]

If one cpu is offline and online later, count_count_start is always 1
and will never equal to ncpus, then, it will loop in the above while and
this cpu will stall.

This adds a count_finish_flag, set it after all cpus are synced and
check it when enter into synchronise_count_slave(), if count_finish_flag
is set, return directly.

BTW, synchronise_count_slave() is called by start_secondary().

Signed-off-by: Wu Zhangjin <zhangjin.wu@windriver.com>