tools: turbostat: make Makefile a bit more capable
The turbostat Makefile was pretty simple, its output would be placed
in the same directory as the source, the install rule had no
concept of a prefix or sysroot, and the default make rules would
be used preventing use of a cross compiler. By making a few minor
changes these limitations are removed while leaving the default
behavior matching what it used to be.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>