tipc: Correction to "Ignore broadcast acknowledgements ..." patch
Revises the checks that were added to TIPC's broadcast link by the
patch "Ignore broadcast acknowledgements that are out-of-range" to
prevent a failed logic assertion from triggering a kernel BUG report.

This revision ensures that any unsent messages in the broadcast link's
transmit queue are properly purged when contact with TIPC's only
neighboring node is lost, thereby restoring the clean up capability
that existed before the new broadcast link checks were added. The checks
are now performed only if TIPC is still in contact with the acknowledging
node. Conversely, if contact with the acknowledging node has been lost
the checks are unnecessary and the broadcast link determines which
messages in the transmit queue need to be cleaned up, based on whether
or not the unsent messages (if any) can still be delivered.

Signed-off-by: Allan Stephens <allan.stephens@windriver.com>