oscb: introduce the OS control block
The OS Control Block (OSCB) is a block of memory than an execution
environment (operating system, executive, hypervisor etc) can use to
describe itself to the world. It is particularily useful for tools to
adapt how they interact with a target.

This change introduces the OSCB and enables it by default. All
fields are initialized at compile time and modules can reference
the OSCB via the variable oscb_header.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>