OMAP: Introduce a user list for each voltage domain instance in the voltage driver.
commit feaf9bb7bb5f119840e197cd907cdd9eaa418f23 from
git://arago-project.org/git/projects/linux-omap3.git

This patch introduces a user list of devices associated with each
voltage domain instance. The user list is implemented using plist
structure with priority node populated with the voltage values.
This patch also adds an API which will take in a device and
requested voltage as parameters, adds the info to the user list
and returns back the maximum voltage requested by all the user
devices. This can be used anytime to get the voltage that the
voltage domain instance can be transitioned into.

Signed-off-by: Thara Gopinath <thara@ti.com>
Integrated-by: Stanley.Miao <stanley.miao@windriver.com>