OMAP: Introduce API to register a device with a voltagedomain
commit 1fd74e20f624b555c8d154059690cd13b5daba94 from
git://arago-project.org/git/projects/linux-omap3.git

This patch adds an API in the voltage layer that
can be used  during omap_device_build to register the built
device with the voltage domain. This API is to be typically called
only once per device during the device registeration. This approach
makes it easy during dvfs to scale all the devices associated with
a voltage domain and then scale the voltage domain.

Signed-off-by: Thara Gopinath <thara@ti.com>
Integrated-by: Stanley.Miao <stanley.miao@windriver.com>