tipc: Correction to "Ensure both nodes recognize loss of contact" patch
Supplements the checks that were added to TIPC's link code by the
patch "Ensure both nodes recognize loss of contact between them" to
prevent problems in re-establishing links at both ends.

1) TIPC now suppresses the sending of (most) link protocol messages
to a neighboring node while it is cleaning up after an earlier loss
of contact with that node; this stops the peer node from prematurely
activating its link endpoint, which would prevent TIPC from later
activating its own end. TIPC still allows outgoing RESET messages to
occur during cleanup, to avoid problems if its own node recognizes
the loss of contact first and tries to notify the peer of the situation.

2) TIPC now recognizes an impending loss of contact with a peer node
as soon as it receives a RESET message on a working link that is the
peer's only link to the node, and ensures that the link protocol
suppression mentioned above goes into effect right away -- that is,
even before its own link endpoints have failed. This is necessary to
ensure correct operation when there are redundant links between the nodes,
since otherwise TIPC would send an ACTIVATE message upon receiving a RESET
on its first link and only begin suppressing when a RESET on its second
link was received, instead of initiating suppression with the first RESET
message as it needs to.

Signed-off-by: Allan Stephens <allan.stephens@windriver.com>