gpu/pvr: SGX-KM Initial SGX driver integration for kernel
referred to: 0fd2d775 of git://gitorious.org/pandroid/kernel-omap.git

This is the first version of SGX driver integration to ensure UI
boot-up. Power management and hwmod modifications will be added
as patches.

Signed-off-by: Hemant Hariyani <hemanthariyani@ti.com>

[Only keep two parts, 1) Makefile and Kconfig, 2) pvr/display
directory, because except pvr/display, all other source files will be
replaced by ones in the Ti Graphics_SDK.4.03.00.01. For files in the
pvr/display, they are for OMAP4 SGX540 Virtual Display functions and
are not in the Graphics_SDK package yet.]

Integrated-by: Hui Wang <Hui.Wang@windriver.com>