gpu/pvr: Modify Makefile to be consistent with Graphics_SDK.4.03.00.01
Source codes in the Graphics_SDK need different macros when building,
so changed macros passed to ccflags.

Original Makefile only support drivers to be built as built-in, so
change it to support drivers to be built both as built-in or as
kernel module.

Original Makefile will generate different objs from Graphics_SDK, so
change Makefile to generate the same name modules as Graphics_SDK.

Signed-off-by: Hui Wang <Hui.Wang@windriver.com>