OMAP: DSS2: FIFI UNDERFLOW issue fixed
commit 4b58d194fd8f5a1c68803eb326e9a71621571696 from
the branch origin/OMAPPSP_03.00.01.06 of
git://arago-project.org/git/projects/linux-omap3.git

In case of 720P with 90 degree rotation, the system reports GFX_FIFO_UNDERFLOW
error which usually happens if DSS DMA is not able to fill the FIFO as per
requirement.
In TRM (section 11.2.6.1.3), where is has been clearly mentioned that,

"To improve the performance on 90/270 degree rotation, split the data access on
write side and not read side."

That means, read should always happen on 0 degree and write should go to
respective rotation view.

Signed-off-by: Vaibhav Hiremath <hvaibhav@ti.com>
Integrated-by: Stanley.Miao <stanley.miao@windriver.com>