mpc85xx: Save scratch registers to thread info instead of using SPRGs.
Extracted from vendor drop QorIQ-DPAA-SDK-V1-20110609-systembuilder.iso
0090-mpc85xx-Save-scratch-registers-to-thread-info-instea.patch

We expect this is actually faster, and we end up needing more space than we
can get from the SPRGs in some instances.  This is also useful when running
as a guest OS - SPRGs4-7 do not have guest versions.

8 slots are allocated in thread_info for this even though we only actually
use 4 of them - this allows space for future code to have more scratch
space (and we know we'll need it for things like hugetlb).

Signed-off-by: Ashish Kalra <Ashish.Kalra@freescale.com>
Signed-off-by: Becky Bruce <beckyb@kernel.crashing.org>
Signed-off-by: Kumar Gala <galak@kernel.crashing.org>