fsl_usdpaa_shmem: reworked from old "fsl_shmem" driver.
Extracted from vendor drop QorIQ-DPAA-SDK-V1-20110609-systembuilder.iso
0192-fsl_usdpaa_shmem-reworked-from-old-fsl_shmem-driver.patch

This the old "fsl_shmem" driver, renamed to "fsl_usdpaa_shmem", with a
variety of updates that make it USDPAA-specific.

The support for a "fsl-shmem" device-tree node specifying the memory
region is removed, as this required the manual coordination of too many
hard-coded constants. (User-space address-handling, the "mem=" boot
parameter to the kernel, the buffer-pool seeding in the device-tree, and
the "TLB1 hack" in the kernel memory handling.) Note also that the
original authors of fsl_shmem, who believed it to be removed for good,
were not keen to see it revived in its existing form.

This change uses a Kconfig setting to trigger an early-boot allocation of
a physically-contiguous memory range, size is also controlled by a
Kconfig setting (default: 64MB).

Also, the old "TLB1 hack" is included and has been updated to match
against the early-boot allocation, rather than assuming hard-coded
addresses. This also means the driver cannot be built as a loadable
module, it must be built into the kernel. The driver's early-boot hook
will attempt to find an unused region of suitably-aligned memory, by
searching from the high-end of the memory range downwards, though due to a
conflict with ramdisk handling this logic intentionally limits itself to
the first 6GB. It also reserves a "tlbcam_index" value to ensure other
uses of TLB1 (HugeTLB, kvm, ...) don't conflict.

The resulting character device "/dev/fsl-usdpaa-shmem" (renamed) supports
a new ioctl() to allow the application to retrieve the physical (DMA)
base-address and size of the memory region. (So this change is inter-
dependent on a matching change in the user-space "dma_mem" driver.)

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>
Signed-off-by: Becky Bruce <beckyb@kernel.crashing.org>
[context adjustment]
Integrated-by: Weiwei Wang <weiwei.wang@windriver.com>