intel_atom_z530: add the rev 1922 / 1.6 Gold version of EMGD
The Embedded Media and Graphics Driver provides support for the GMA 500 /
GMA 600 GPUs in Intel Atom Z5xx / Z6xx / E6xx processors.  These GPUs
are based on the PowerVR SGX cores from Imagination Technologies.
This kernel driver goes with a set of proprietary X modules and related
libraries in user space, provided through the intel-emgd template.  The
rev 1922 version of this driver is the first one that has a license
that is compatible with GPL v2 and is redistributable under a production
license.  The driver has been tested on the Portwell Nano 8044 ITX board.

Signed-off-by: Donn Seeley <donn.seeley@windriver.com>
Signed-off-by: Hui Wang <Hui.Wang@windriver.com>