powerpc/mm: 64-bit: Don't use MMU_FTR_USE_PAIRED_MAS in tlb handler
Extracted from vendor drop QorIQ-DPAA-SDK-V1-20110609-systembuilder.iso
0190-powerpc-mm-64-bit-Don-t-use-MMU_FTR_USE_PAIRED_MAS-i.patch

It is mandatory in the architecture for 64-bit book3e, so unless we know
of a specific 64-bit book3e chip that doesn't implement them, don't
bother with the checks.  The nops still consume decode bandwidth, and
appear to be slowing us down.

Signed-off-by: Scott Wood <scottwood@freescale.com>
Signed-off-by: Kumar Gala <galak@kernel.crashing.org>