dpaa_eth: Handle 10G MDIO read errors
Extracted from vendor drop QorIQ-DPAA-SDK-V1-20110609-systembuilder.iso
0157-dpaa_eth-Handle-10G-MDIO-read-errors.patch

When the 10G MDIO bus encounters an unresponsive PHY or device, it
sets a flag indicating a read error, and it leaves the data field
as it was. This means that reads of non-existent PHYs/devices returns
the previously read value from the last read device. Also, if we are
asked to read in a Clause-22 fashion, we were attempting to read from
device -1, which might be interpretted as device 31.  For both cases,
we now return 0xffff, which is often interpretted by upper layers of
code as a failed read.

Signed-off-by: Andy Fleming <afleming@freescale.com>