dpaa_eth: Rearrange 1588 port initialization
Extracted from vendor drop QorIQ-DPAA-SDK-V1-20110609-systembuilder.iso
0158-dpaa_eth-Rearrange-1588-port-initialization.patch

The 1588 port initialization will only be done on Linux netdevices
with MACs.  The code stores the timer information in the netdev
priv structure, which only exists for Linux network interfaces.
However, the port initialization may be done for proxy interfaces
(ie, MACs which are initialized by the driver on behalf of another
partition), so we can't access "priv" in that code.  The fix is to
move the access up to where we know we have a net device and a MAC.

Signed-off-by: Andy Fleming <afleming@freescale.com>