markers-list-from-marker-hash-table
Markers - list from marker hash table

The marker hash table contains the correct list of markers which are enabled,
even the ones which does have a marker instance in any module. This is the table
we should dump in the trace, not the markers section, because otherwise we can
miss a marker if it's connected before its module is loaded.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@polymtl.ca>