wrhv/p4080: disable the ethernet which can't be used
The P4080 SoC contains 10 total ethernet MACs, but they can't all be
used at the same time for the reasons of pin multiplexing. Currently
we use SerDes protocol 0x10 on P4080. It means that only FM1@DTSEC2,
FM2@DTSEC1, FM2@DTSEC2 and FM2@TGEC1 can be used. We already comment
out FM1@DTSEC1, FM1@DTSEC3, FM1@DTSEC4 and FM1@TGEC1 device nodes in
DTS, but leave FM2@DTSEC3 and FM2@DTSEC4 still here. And the DPAA
driver will also allocate resources for these two ethernets which
can't not be used. Then some resources such as the number of tasks on
FMAN will be exhausted before initializing the FM2@TGEC1, and cause
it malfunction.

Instead of commenting out the ethernet device node, we just mark them
as "disabled" in DTS. This is also what the u-boot do for the native
BSP.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>