dpaa_eth: Support checksumming on VLAN packets
Extracted from vendor drop QorIQ-DPAA-SDK-V1-20110609-systembuilder.iso
0160-dpaa_eth-Support-checksumming-on-VLAN-packets.patch

In order to do this, we have to propagate the netdev features
to the vlan device. Also, we need to look for the IP type in the
encapsulated header.

While changing that code, clean up the printks to use cpu_netdev_alert(),
and use the *_offset() helpers to determine the offsets.

Signed-off-by: Andy Fleming <afleming@freescale.com>