smc91x: fix kgdboe connection failure
Ensure locking is preemption safe. Add a small delay in smc_poll_controller()
to ensure that when running a target with smc91x ethernet as a Qemu guest and
using KGDBoE, that Qemu's interrupt handling loop in cpu_exec() can exit,
allowing Qemu to process new network packets. Without this delay the network
packets will not be process and it will appear as if the connection between
gdb and the target has failed, errors such as

    (gdb) n
    Ignoring packet error, continuing...
    Reply contains invalid hex digit 116

With this delay qemu will still use large cpu load and exhibit some
instability due to this but overall it will remain responsive.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>
Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>