dpaa_eth: Store dpa_bp counts in percpu_priv
Extracted from vendor drop QorIQ-DPAA-SDK-V1-20110609-systembuilder.iso
0162-dpaa_eth-Store-dpa_bp-counts-in-percpu_priv.patch

Accessing the per_cpu area is somewhat costly, and even more so on
64-bit. We have a percpu struct for each interface, and a percpu counter
for each buffer pool. This meant that we were going through the percpu
access process twice. In order to improve this, we put a pointer to the
buffer pool counter in the percpu_priv structure for each interface.

Signed-off-by: Andy Fleming <afleming@freescale.com>