kgdboc: fix lost characters with scheduled breakpoint
When using the rx_poll api to snoop a tty session for the '$'
character or a control-c, a break point is scheduled when either
character is detected.  All of the data must then go into the queue up
until kgdb is fully entered.

Prior to this patch these characters would be lost and as a result the
debugger connection sequences for gdb would end up timing out or have
a retry or sometimes occasionally complete failure to connect.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>