wrhv/dpa: release the QMAN/BMAN resource when kernel restart
In guest OS, the QMAN/BMAN doesn't really be reset when the
kernel restart. So we should release the buffers in BMAN which are
created by dpa driver, and also destroy the FQs in QMAN which are
used for tx/rx queue. Otherwise after reboot the BMAN will include
stale buffers, and the FQ which is still in running state will be
initialized again.  This will make these FQs nonoperative.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
[rebase it onto QorIQ-DPAA-SDK-V1-20110609-systembuilder.iso]
Integrated-by: Weiwei Wang <weiwei.wang@windriver.com>