Q/BMAN: set IRQF_NOBALANCING for the portal irq handler
The Q/BMAN portal is bound to a specific cpu. It make no sense
to migrate the portal irq to orther cpu when a cpu is down. And the
portal irq handler also doesn't support reentry. That mean we can't
handle two portal irq on a single cpu. So set the portal irq with
IQF_NOBALANCING flag to avoid irq migration or setting affinity
by userspace.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
[rebase it onto QorIQ-DPAA-SDK-V1-20110609-systembuilder.iso]
Integrated-by: Weiwei Wang <weiwei.wang@windriver.com>