net/dpa: only use one buffer pool even when static buffer pools are used
Currently the DPA driver only supports one buffer pool, and all the
ethernet will use it. This works fine when no static buffer pool is
used. But when we assign different static buffer pool to each ethernet,
the dpa driver will allocate a buffer pool for each ethernet, and use
the static bpid to initialize the port register. But it doesn't really
allocate the buffers for these pools except the first one. This will
definitely cause the others ethernet malfunction. So we should not
allocate new buffer pools after one has already been created even when
we use static buffer pool for this ethernet.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
[rebase it onto QorIQ-DPAA-SDK-V1-20110609-systembuilder.iso]
Integrated-by: Weiwei Wang <weiwei.wang@windriver.com>