MMC/SD: power on before detecting card
The driver will not get the right status when using the callback
function 'get_cd' to check if the card is present because the
power is not on. Therefore, the controller must be power on
before detecting the card status.

Extracted from vendor drop QorIQ-NONDPAA-SDK-V1-20110429_ltib.iso
linux-2.6.35-qoriq-MMC-SD-power-on-before-detecting-card.patch.

Signed-off-by: Jerry Huang <Chang-Ming.Huang@freescale.com>
Integrated-by: Xufeng Zhang <xufeng.zhang@windriver.com>