wrhv: powerpc: introduce powerpc specifics of WR guest
These powerpc specific additons are the Linux specific
additions that were not a part of the reference VBI
implementation.

o support for e500 and e500mc guest
[
- create e500 common config option
- implement tlb invalidate function based tlbilx/tlbivax.
- paravirtualize TLB cache flushing
- set_context behaves different on e500/e500mc
  For e500mc set_context uses the PID register to point to
  the current process.  Where as the standard e500 needs to
  copy the kernel page tables from the init process to the
  current process.

- set proper permissions on kernel pages
  The hypervisor needs the TLB pages to be marked as:
  SX/SW/SR/UW/UR/UX
]

o enable SPE support for E500
[
Add SPE support for E500 core. Currently in hypervisor the
debug exception is defined by default. So we have to add a
dummy handler in linux kernel in order to create the correct
interrupt vector offset as the hypervisor.

When support SPE we introduce a dummy handler in linux kernel in order
to create the correct SPE interrupt vector offset as the hypervisor. But
actually that should be used for debug exception is defined by guest OS.
]

o support for SMP
[
Provide basic e500 SMP support which will be shared by all
non e500mc hypervisor guests.

Our implementation differs from upstream in that it does
not make use of the MPIC for IPIs, rather it uses the VIOAPIC
as its IPI communication mechanism.
]

o use kmalloc instead of vmalloc for modules
[
Replace vmalloc(...) with kmalloc(..., GFP_KERNEL) for PPC.
]

wrhv: powerpc: misc fixes
[
wrhv/e500: tlb_cam fixes for paravirtualized kernels

The following commit in 2.6.34.1 adjusts the tlb cam code
for e500 based boards.

commit b8504b8d34bf036b952c3f3b0f5cd5864d5f42e5
Author: Kumar Gala <galak@kernel.crashing.org>
Date:   Thu May 13 14:38:21 2010 -0500

    powerpc/fsl-booke: Move loadcam_entry back to asm code to fix SMP ftrace

    commit 78f622377f7d31d988db350a43c5689dd5f31876 upstream.

    When we build with ftrace enabled its possible that loadcam_entry would
    have used the stack pointer (even though the code doesn't need it).  We
    call loadcam_entry in __secondary_start before the stack is setup.  To
    ensure that loadcam_entry doesn't use the stack pointer the easiest
    solution is to just have it in asm code.

    Signed-off-by: Kumar Gala <galak@kernel.crashing.org>
    Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>

We need to modify the paravirtualized code to declare and define
the functions that have been relocated.
]

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>
Signed-off-by: Tiejun Chen <tiejun.chen@windriver.com>
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
Signed-off-by: Jeremy McNicoll <jeremy.mcnicoll@windriver.com>
Signed-off-by: Liang Li <liang.li@windriver.com>