gpu/pvr: minor fixes for EMGD code
o emgd need drm_helper so select it by default

o fix the wrong position of the mutex definition
[
The wrong code piece is #ifdef'd by an undefined MACRO, the MACRO is
for 'debugging purpose', which is not enabled in Makefile. When enable
some debugging MACRO in the Makefile, build error will occur.
]

Signed-off-by: Liang Li <liang.li@windriver.com>