net/dpa: set CONFIG_FSL_FMAN default value as n
Currently CONFIG_FSL_FMAN is enabled by default, and the
result is CONFIG_FSL_FMAN is always y in all boards. it is not
right(even no harm), so here set CONFIG_FSL_FMAN to n by default.
CONFIG_FSL_FMAN needs to be turned on explicitly in board cfg.

Signed-off-by: Weiwei Wang <weiwei.wang@windriver.com>