powerpc/85xx: Save scratch registers to thread info instead of using SPRGs
commit 1325a684b553d4b5c41ae0482f8991b43f945746 upstream.

We expect this is actually faster, and we end up needing more space than we
can get from the SPRGs in some instances.  This is also useful when running
as a guest OS - SPRGs4-7 do not have guest versions.

8 slots are allocated in thread_info for this even though we only actually
use 4 of them - this allows space for future code to have more scratch
space (and we know we'll need it for things like hugetlb).

Here we should port this commit to support Wind River p4080 GOS.

Signed-off-by: Tiejun Chen <tiejun.chen@windriver.com>