gianfar: add support for wake-on-packet
The wakeup-user-defined packet is ARP request packet or
unicast packet to board.

Gianfar driver receives the wakeup-user-defined packet to
rx queue#1, the normal packet to rx queue#0.

We set filter table to filer all of packets except the
wake up user defined packet to it.

The rule of the filter table as below
if (arp request to local ip address)
       accept it to queue#1
elif (unicast packet to local mac address)
       accept it to queue#1
else
       reject it
endif
* The local ip/mac address is the ethernet ip/mac address
of the baord

The receive part of eTSEC still continues work to
get the wake up user defined packet and generates interrupt to
wake up the system.

Extracted from vendor drop QorIQ-NONDPAA-SDK-V1-20110429_ltib.iso
linux-2.6.35-qoriq-gianfar-add-support-for-wake-on-packet.patch.

Signed-off-by: Dave Liu <daveliu@freescale.com>
Signed-off-by: Jin Qing <b24347@freescale.com>
Signed-off-by: Li Yang <leoli@freescale.com>
Integrated-by: Xufeng Zhang <xufeng.zhang@windriver.com>