gianfar: Change the TX SCHED from PRIO to WRRS
1. With PRIO scheduling, priority is always given to Q0
and when there are multiple active queues, we will notice
a performance degradation for larger byte packets.
To resolve the performance issue for larger byte packets use
WRRS with 0xFF as weight value for each queue.
Also note that for proper functioning of WRRS, the weight for
each queue should be greater than the max packet size that is
supported by default. Hence we used value 0xFF (max value) for
each queue.

2. Also made minor changes to the poll function to reduce the number
of iterations when multiple queues are active.

Extracted from vendor drop QorIQ-NONDPAA-SDK-V1-20110429_ltib.iso
linux-2.6.35-qoriq-gianfar-Change-the-TX-SCHED-from-PRIO-t.patch.

Signed-off-by: Sandeep Gopalpet <sandeep.kumar@freescale.com>
Signed-off-by: Jiajun Wu <b06378@freescale.com>
Integrated-by: Xufeng Zhang <xufeng.zhang@windriver.com>