Bug fix for system crashes on running PTPd when PTP timer node is not defined
When the IEEE1588 node is not defined in the DTS file, and PTPD application
calls 1588_ioctl the system crashes due to null pointer access. The patch adds
a check if ptptimer device is present in the system before calling 1588_ioctl.

Extracted from vendor drop QorIQ-NONDPAA-SDK-V1-20110429_ltib.iso
linux-2.6.35-qoriq-Bug-fix-for-System-crashes-on-running-P.patch.

Signed-off-by: P.V.Suresh <pala@freescale.com>
Integrated-by: Xufeng Zhang <xufeng.zhang@windriver.com>