ARM:CNS3xxx:Migrate PHYS_OFFSET to 0x20000000
On Econa chip, 0x20000000 is base address of DDR2 SDRAM, the size of
this region is 1G byte. 0x00000000 is used for alias memory, which is
a shadow of RAM after remapped in uboot, the size of this region is
256M byte.

To support system with SDRAM more than 256M byte, it needs to move the
PHYS_OFFSET to 0x20000000.

In addition to change the definition of PHYS_OFFSET, the patch update
a few symbol definitions related to physical address.

Signed-off-by: Jiang Lu <lu.jiang@windriver.com>