wrhv/e500: stub emulation for TLB1CFG
We don't have the read permission to TLB1CFG register in GOS. So we
will get the following panic when reading it.

Kernel panic - not syncing: Unsupported read for spr 0x2b1 in
smp_store_cpu_info+0x40/0x64

Call Trace:
[cf81fea0] [c0007d38] show_stack+0x44/0x160 (unreliable)
[cf81fed0] [c05a2cec] panic+0x12c/0x1a4
[cf81ff20] [c00174dc] wrhv_mfspr+0x40/0xd4
[cf81ff30] [c05a6534] smp_store_cpu_info+0x40/0x64
[cf81ff50] [c06e5f88] smp_prepare_cpus+0x38/0x138
[cf81ff80] [c06e0264] kernel_init+0x70/0x278
[cf81fff0] [c0011d8c] original_kernel_thread+0x4c/0x68

Since we never use the contents of this register in GOS, add a stub
emulation to return 0 when we are reading it. The reason we choose to
return 0 here is that it is more easy to cause a misfunction if this
value is used unintentionally by some code in GOS.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>