PPC_BOOK3E/kexec: implement pseudo MMU real mode
PPC64 Kexec always load Image to so-called real mode that turn off mmu
but its impossible for book3E. So we have to setup the initial TLB for each
core to map v:0 to p:0 as 1:1, this would make sure on Book3E we can access
everything from virtual address 0x0. Note this current implementation
assume that 1G is enough for kexec.

Signed-off-by: Tiejun Chen <tiejun.chen@windriver.com>