qbman: only create affine portal for online cpu
Currently the qbman driver will try to create affine portal for
all the present cpu. If the kernel finds that the cpu is not online,
it will redirects this portal's irq to an online cpu at random.
But the qbman's irq handler only supports one affine portal for
each cpu. Then the cpu which is assigned two portal's irq will
get stuck. And it also makes no sense to create an affine portal
for an offline cpu.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>