drivers:ramconsole: fix not proper references in other arches
When compiling ram console code we were getting build errors due to missing
functions. These functions are only available when specific kernel configs are
set, so block out the use of these functions based on kernel config settings.
When in place these omissions will not affect the functionality of the code.

Signed-off-by: Zumeng Chen <zumeng.chen@winriver.com>